refactor: DiaryItemManager에 TextViewManager의 기능을 통합
